Templatize ApproxEqualFailure and assert_approx_equal
ApproxEqualFailure becomes a class template ApproxEqualFailure<T>.
Its constructor uses string_from(T) for formatting and detects NaN via
the IEEE 754 self-comparison trick (v != v), the same approach used in
approx_equal — so no type-specific code, no hard-coded double, no
string matching.

Shell::assert_approx_equal becomes a non-virtual template method that
constructs ApproxEqualFailure<T> directly. check_approx<T> passes its
T values straight through, with no cast and no dispatch helper. The
virtual seam for overriding is add_failure, as with every other
assertion.
