+++ include/mutiny/test/Shell.hpp
@@ -11,6 +11,7 @@
 #ifndef INCLUDED_MUTINY_TEST_SHELL_HPP
 #define INCLUDED_MUTINY_TEST_SHELL_HPP
 
+#include "mutiny/test/Failure.hpp"
 #include "mutiny/test/Terminator.hpp"
 
 #include "mutiny/String.hpp"
@@ -246,26 +247,27 @@ public:
       size_t line_number,
       const Terminator& test_terminator = get_current_test_terminator()
   );
-  /** @brief Macro backend: assert two doubles are equal within @p threshold. */
-  virtual void assert_approx_equal(
-      double expected,
-      double actual,
-      double threshold,
+  /**
+   * @brief Macro backend: assert two values are equal within @p threshold.
+   *
+   * @tparam T  Numeric type; must have a @ref mu::tiny::string_from() overload.
+   */
+  template<typename T>
+  void assert_approx_equal(
+      T expected,
+      T actual,
+      T threshold,
       const char* text,
       const char* file_name,
       size_t line_number,
       const Terminator& test_terminator = get_current_test_terminator()
-  );
-  /** @brief Macro backend: assert two floats are equal within @p threshold. */
-  virtual void assert_approx_equal(
-      float expected,
-      float actual,
-      float threshold,
-      const char* text,
-      const char* file_name,
-      size_t line_number,
-      const Terminator& test_terminator = get_current_test_terminator()
-  );
+  )
+  {
+    add_failure(
+        ApproxEqualFailure<T>(this, file_name, line_number, expected, actual, threshold, text)
+    );
+    test_terminator.exit_current_test();
+  }
   /** @brief Macro backend: generic equality failure with pre-formatted strings.
    */
   virtual void assert_equals(
@@ -541,48 +543,6 @@ void check_enum_equal(
 }
 
 /**
- * @brief Failure-reporting dispatch for @ref check_approx.
- *
- * Primary template: casts @p T to @c double so that integral types produce an
- * unambiguous overload call and the conversion is explicit rather than silent.
- * Explicitly specialised for @c float (see below) to avoid promoting float
- * values to double.
- */
-template<typename T>
-void check_approx_fail(
-    T expected,
-    T actual,
-    T threshold,
-    const char* text,
-    const char* file,
-    size_t line
-)
-{
-  Shell::get_current()->assert_approx_equal(
-      static_cast<double>(expected),
-      static_cast<double>(actual),
-      static_cast<double>(threshold),
-      text,
-      file,
-      line
-  );
-}
-
-/** @brief @c float specialisation: passes values as @c float, no promotion. */
-template<>
-inline void check_approx_fail<float>(
-    float expected,
-    float actual,
-    float threshold,
-    const char* text,
-    const char* file,
-    size_t line
-)
-{
-  Shell::get_current()->assert_approx_equal(expected, actual, threshold, text, file, line);
-}
-
-/**
  * @brief Implementation helper for CHECK_APPROX.
  *
  * All three operands share the same type @p T so mismatched-type calls produce
@@ -607,7 +567,7 @@ void check_approx(
 )
 {
   if (!approx_equal(expected, actual, threshold)) {
-    check_approx_fail(expected, actual, threshold, text, file, line);
+    Shell::get_current()->assert_approx_equal(expected, actual, threshold, text, file, line);
   } else {
     Shell::get_current()->count_check();
   }
